requirements.txt
pyspark==3.5.0
pandas==2.0.3
numpy==1.24.3

kafka-python==2.0.2

scikit-learn==1.3.0
scipy==1.11.4

python-dateutil==2.8.2
pytz==2023.3

matplotlib==3.7.2
seaborn==0.12.2

tqdm==4.66.1

 colorlog==6.7.0